Added a damage system & added it to the bomberman state.

+++ include/Client/HealthComponent.hpp
@@ -12,7 +12,7 @@ namespace Client
             HealthComponent(int health);
             virtual ~HealthComponent();
 
-            int health = 1;
+            float health = 1;
     };
 }
